requirements.txt
absl-py==0.10.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.11.1
black==21.12b0
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
dataclasses==0.8
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
entrypoints==0.3
etils==0.6.0
executing==0.8.2
filelock==3.7.1
flatbuffers==2.0
fonttools==4.29.1
frozenlist==1.2.0
gast==0.4.0
gensim==4.2.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
grpcio==1.43.0
h5py==3.1.0
huggingface-hub==0.7.0
idna==3.3
imageio==2.19.2
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.1
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
librosa==0.8.1
llvmlite==0.38.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
munkres==1.1.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.8
nltk==3.7
notebook==6.4.7
numba==0.55.1
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.3.1
platformdirs==2.4.1
plotly==5.8.0
pooch==1.6.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2021.3
pyu2f==0.1.5
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.8
sacremoses==0.0.43
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==60.5.0
six==1.15.0
smart-open==6.0.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
stack-data==0.1.4
statsmodels==0.13.2
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-macos==2.7.0
tensorflow-metadata==1.8.0
tensorflow-metal==0.3.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tokenizers==0.12.1
toml==0.10.2
tomli==1.2.3
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.19.2
typing_extensions==4.0.1
unicodedata2==14.0.0
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.35.1
widgetsnbextension==3.5.2
wordcloud==1.8.1
wrapt==1.12.1
yarl==1.7.2
zipp==3.7.0
